upstream merge & updates to tests
