+++ .travis.yml
@@ -65,7 +65,7 @@ install:
   - python condaci.py setup
 
 script:
-  - pytest --cov=autocnet
+  - pytest autocnet functional_tests
 
 after_success:
   # Upload to anaconda and push to coveralls
