+++ autocnet/control/tests/test_control.py
@@ -1,107 +1,13 @@
+from unittest.mock import MagicMock
+import geopandas as gpd
+import pandas as pd
+from shapely.geometry import Polygon
+
+
 import os
 import sys
-import itertools
-
-from time import gmtime, strftime
-
-import pytest
-
-from unittest.mock import Mock, MagicMock
-
-
-import numpy as np
-import pandas as pd
-
-sys.path.insert(0, os.path.abspath('..'))
-
-from autocnet.control import control
-
-
-@pytest.fixture
-def controlnetwork_data():
-    df = pd.DataFrame([[0, 0.0, 0.0, (0.0, 1.0), 0, 0.0, 0.0],
-                             [0, 1.0, 0.0, (0.0, 1.0), 0, 0.0, 0.0],
-                             [1, 0.0, 1.0, (0.0, 1.0), 1, 0.0, 0.0],
-                             [1, 1.0, 1.0, (0.0, 1.0), 1, 0.0, 0.0],
-                             [2, 0.0, 2.0, (0.0, 1.0), 2, 0.0, 0.0],
-                             [2, 1.0, 2.0, (0.0, 1.0), 2, 0.0, 0.0],
-                             [3, 0.0, 3.0, (0.0, 1.0), 3, 0.0, 0.0],
-                             [3, 1.0, 3.0, (0.0, 1.0), 3, 0.0, 0.0],
-                             [4, 0.0, 4.0, (0.0, 1.0), 4, 0.0, 0.0],
-                             [4, 1.0, 4.0, (0.0, 1.0), 4, 0.0, 0.0],
-                             [5, 0.0, 5.0, (0.0, 1.0), 5, 0.0, 0.0],
-                             [5, 1.0, 5.0, (0.0, 1.0), 5, 0.0, 0.0],
-                             [6, 0.0, 6.0, (0.0, 1.0), 6, 0.0, 0.0],
-                             [6, 1.0, 6.0, (0.0, 1.0), 6, 0.0, 0.0],
-                             [7, 0.0, 7.0, (0.0, 1.0), 7, 0.0, 0.0],
-                             [7, 1.0, 7.0, (0.0, 1.0), 7, 0.0, 0.0],
-                             [0, 2.0, 0.0, (0.0, 2.0), 0, 0.0, 0.0],
-                             [1, 2.0, 1.0, (0.0, 2.0), 1, 0.0, 0.0],
-                             [2, 2.0, 2.0, (0.0, 2.0), 2, 0.0, 0.0],
-                             [3, 2.0, 3.0, (0.0, 2.0), 3, 0.0, 0.0],
-                             [4, 2.0, 4.0, (0.0, 2.0), 4, 0.0, 0.0],
-                             [5, 2.0, 5.0, (0.0, 2.0), 5, 0.0, 0.0],
-                             [8, 0.0, 8.0, (0.0, 2.0), 6, 0.0, 0.0],
-                             [8, 2.0, 8.0, (0.0, 2.0), 6, 0.0, 0.0],
-                             [9, 0.0, 9.0, (0.0, 2.0), 7, 0.0, 0.0],
-                             [9, 2.0, 9.0, (0.0, 2.0), 7, 0.0, 0.0],
-                             [10, 1.0, 8.0, (1.0, 2.0), 6, 0.0, 0.0],
-                             [10, 2.0, 6.0, (1.0, 2.0), 6, 0.0, 0.0],
-                             [11, 1.0, 9.0, (1.0, 2.0), 7, 0.0, 0.0],
-                             [11, 2.0, 7.0, (1.0, 2.0), 7, 0.0, 0.0]],
-                            columns=['point_id', 'image_index', 'keypoint_index',
-                                     'edge', 'match_idx', 'x', 'y'])
-
-    df.index.name = 'measure_id'
-
-    #Fix types
-    df['point_id'] = df['point_id'].astype(object)
-    df['match_idx'] = df['match_idx'].astype(object)
-
-    return df
-
-@pytest.fixture
-def candidategraph():
-    edges = [(0,1), (0,2), (1,2)]
-
-
-    match_indices = [([0,1,2,3,4,5,6,7], [0,1,2,3,4,5,6,7]),
-                     ([0,1,2,3,4,5,8,9], [0,1,2,3,4,5,8,9]),
-                     ([0,1,2,3,4,5,8,9], [0,1,2,3,4,5,6,7])]
-
-    matches = []
-    for i, e in enumerate(edges):
-        c = match_indices[i]
-        source_image = np.repeat(e[0], 8)
-        destin_image = np.repeat(e[1], 8)
-        coords = np.zeros(8)
-        data = np.vstack((source_image, c[0], destin_image, c[1],
-                          coords, coords, coords, coords)).T
-        matches_df = pd.DataFrame(data, columns=['source_image', 'source_idx', 'destination_image', 'destination_idx',
-                                                 'source_x', 'source_y', 'destination_x', 'destination_y'])
-        matches.append(matches_df)
-    # Mock in the candidate graph
-    cg = MagicMock()
-    cg.get_matches = MagicMock(return_value=matches)
-
-    return cg
-
-@pytest.fixture()
-def controlnetwork(controlnetwork_data):
-    cn = control.ControlNetwork()
-    cn.data = controlnetwork_data
-    # Patching data this way does NOT update the internal _measure_id and _point_id attributes
-    return cn
-
-@pytest.fixture()
-def bad_controlnetwork(controlnetwork_data):
-    cn = control.ControlNetwork()
-    cn.data = controlnetwork_data
-    # Since the data is being patched in, fix the measure counter
-    cn._measure_id = len(cn.data) + 1
-    # Add a duplicate measure in image 0 to point 0
-    cn.add_measure((0,11), (0,1), 2, [1,1], point_id=0)
-    return cn
+sys.path.insert(0, '..')
+from .. import control
 
 def test_fromcandidategraph(candidategraph, controlnetwork_data):#, controlnetwork):
     matches = candidategraph.get_matches()
@@ -136,5 +42,27 @@ def test_bad_validate_points(bad_controlnetwork):
     assert bad_controlnetwork.validate_points().iloc[0] == False
     assert bad_controlnetwork.validate_points().iloc[1:].all()
 
-def test_potential_overlap(controlnetwork):
-    pass
+def test_identify_potential_overlaps(controlnetwork, candidategraph):
+    res = control.identify_potential_overlaps(candidategraph,
+                                              controlnetwork,
+                                              overlap=False)
+
+    assert res.equals(pd.Series([(2,), (2,),
+                                 (1,), (1,),
+                                 (0,), (0,)],
+                                 index=[6,7,8,9,10,11]))
+
+def test_potential_overlap(controlnetwork, candidategraph):
+    # Patch in an intersection check so that all points intersect all geoms
+    candidategraph.create_node_subgraph = MagicMock(return_value=candidategraph)
+    coords = [(-1., -1.), (-1., 1.), (1., 1.), (1., -1.), (-1., -1.)]
+    poly = gpd.GeoSeries(Polygon(coords))
+    candidategraph.compute_intersection = MagicMock(return_value=(poly, 0))
+    res = control.identify_potential_overlaps(candidategraph,
+                                              controlnetwork,
+                                              overlap=True)
+
+    assert res.equals(pd.Series([(2,), (2,),
+                                 (1,), (1,),
+                                 (0,), (0,)],
+                                 index=[6,7,8,9,10,11]))
