+++ autocnet/graph/node.py
@@ -66,8 +66,6 @@ class Node(dict, MutableMapping):
         self['node_id'] = node_id
         self['hash'] = image_name
         self._mask_arrays = {}
-        self.point_to_correspondence = defaultdict(set)
-        self.point_to_correspondence_df = None
         self.descriptors = None
         self.keypoints = pd.DataFrame()
         self.masks = pd.DataFrame()
@@ -116,26 +114,13 @@ class Node(dict, MutableMapping):
         for k, v in d.items():
             if isinstance(v, pd.DataFrame):
                 if not v.equals(o[k]):
+                    print('NODE', k)
                     eq = False
             elif isinstance(v, np.ndarray):
                 if not v.all() == o[k].all():
+                    print('NODE', k)
                     eq = False
         return eq
-    """
-    def __getitem__(self, item):
-        attribute_dict = {'image_name': self['image_name'],
-                          'image_path': self['image_path'],
-                          'geodata': self.geodata,
-                          'keypoints': self.keypoints,
-                          'nkeypoints': self.nkeypoints,
-                          'descriptors': self.descriptors,
-                          'masks': self.masks,
-                          'isis_serial': self.isis_serial}
-        if item in attribute_dict.keys():
-            return attribute_dict[item]
-        else:
-            return super(Node, self).__getitem__(item)
-    """
 
     @property
     def geodata(self):
@@ -146,31 +131,6 @@ class Node(dict, MutableMapping):
             return self._geodata
         else:
             return None
-
-    """    @property
-    def masks(self):
-        mask_lookup = {'suppression': 'suppression'}
-
-        if self.keypoints is None:
-            warnings.warn('Keypoints have not been extracted')
-            return
-
-        if not hasattr(self, '_masks'):
-            self._masks = pd.DataFrame(index=self.keypoints.index)
-
-        # If the mask is coming form another object that tracks
-        # state, dynamically draw the mask from the object.
-        for c in self._masks.columns:
-            if c in mask_lookup:
-                self._masks[c] = getattr(self, mask_lookup[c]).mask
-        return self._masks
-
-    @masks.setter
-    def masks(self, v):
-        column_name = v[0]
-        boolean_mask = v[1]
-        self.masks[column_name] = boolean_mask
-    """
 
     @property
     def footprint(self):
@@ -460,108 +420,6 @@ class Node(dict, MutableMapping):
 
         io_keypoints.to_npy(self.keypoints, self.descriptors,
                             out_path)
-
-    def group_correspondences(self, cg, *args, deepen=False, **kwargs):
-        """
-
-        Parameters
-        ----------
-        cg : object
-             The graph object this node is a member of
-
-        deepen : bool
-                 If True, attempt to punch matches through to all incident edges.  Default: False
-        """
-        node = self['node_id']
-        # Get the edges incident to the current node
-        incident_edges = set(cg.edges(node)).intersection(set(cg.edges()))
-
-        # If this node is free floating, ignore it.
-        if not incident_edges:
-             # TODO: Add dangling correspondences to control network anyway.  Subgraphs handle this segmentation if req.
-            return
-
-        try:
-            clean_keys = kwargs['clean_keys']
-        except:
-            clean_keys = []
-
-        # Grab all the incident edge matches and concatenate into a group match set.
-        # All share the same source node
-        edge_matches = []
-        for e in incident_edges:
-            edge = cg[e[0]][e[1]]
-            matches, mask = edge.clean(clean_keys=clean_keys)
-            # Add a depth mask that initially mirrors the fundamental mask
-            edge_matches.append(matches)
-        d = pd.concat(edge_matches)
-
-        # Counter for point identifiers
-        pid = 0
-
-        # Iterate through all of the correspondences and attempt to add additional correspondences using
-        # the epipolar constraint
-        for idx, g in d.groupby('source_idx'):
-            # Pull the source index to be used as the search
-            source_idx = g['source_idx'].values[0]
-
-            # Add the point object onto the node
-            point = Point(pid)
-            #print(g[['source_image', 'destination_image']])
-            covered_edges = list(map(tuple, g[['source_image', 'destination_image']].values))
-            s = g['source_image'].iat[0]
-            d = g['destination_image'].iat[0]
-            # The reference edge that we are deepening with
-            ab = cg.edge[covered_edges[0][0]][covered_edges[0][1]]
-
-            # Get the coordinates of the search correspondence
-            ab_keypoints = ab.source.get_keypoint_coordinates(index=g['source_idx'])
-            ab_x = None
-
-            for j, (r_idx, r) in enumerate(g.iterrows()):
-                kp = ab_keypoints.iloc[j].values
-
-                # Homogenize the coord used for epipolar projection
-                if ab_x is None:
-                    ab_x = np.array([kp[0], kp[1], 1.])
-
-                kpd = ab.destination.get_keypoint_coordinates(index=g['destination_idx']).values[0]
-                # Add the existing source and destination correspondences
-                self.point_to_correspondence[point].add((r['source_image'],
-                                                                  Correspondence(r['source_idx'],
-                                                                                 kp[0],
-                                                                                 kp[1],
-                                                                                 serial=self.isis_serial)))
-                self.point_to_correspondence[point].add((r['destination_image'],
-                                                                  Correspondence(r['destination_idx'],
-                                                                                 kpd[0],
-                                                                                 kpd[1],
-                                                                                 serial=cg.node[r['destination_image']].isis_serial)))
-
-            # If the user wants to punch correspondences through
-            if deepen:
-                search_edges = incident_edges.difference(set(covered_edges))
-                for search_edge in search_edges:
-                    bc = cg.edge[search_edge[0]][search_edge[1]]
-                    coords, idx = deepen_correspondences(ab_x, bc, source_idx)
-
-                    if coords is not None:
-                        cg.node[node].point_to_correspondence[point].add((search_edge[1],
-                                                                          Correspondence(idx,
-                                                                                         coords[0],
-                                                                                         coords[1],
-                                                                                         serial=cg.node[search_edge[1]].isis_serial)))
-
-            pid += 1
-
-        # Convert the dict to a dataframe
-        data = []
-        for k, measures in self.point_to_correspondence.items():
-            for image_id, m in measures:
-                data.append((k.point_id, k.point_type, m.serial, m.measure_type, m.x, m.y, image_id))
-
-        columns = ['point_id', 'point_type', 'serialnumber', 'measure_type', 'x', 'y', 'node_id']
-        self.point_to_correspondence_df = pd.DataFrame(data, columns=columns)
 
     def coverage_ratio(self, clean_keys=[]):
         """
