+++ tests/test_save_load.py
@@ -1,0 +1,11 @@
+from autocnet.examples import get_path
+from autocnet.graph.network import CandidateGraph
+from autocnet.io.network import load
+
+def test_save_project(tmpdir, candidategraph):
+    path = tmpdir.join('prject.proj')
+    candidategraph.save(path.strpath)
+
+    candidategraph2 = load(path.strpath)
+
+    assert candidategraph == candidategraph2
