+++ tests/test_three_image.py
